requirements.txt
# TensorFlow and Core ML Libraries
tensorflow>=2.10.0,<2.16.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Data Processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration
PyYAML>=5.4.0

# MediaPipe for pose extraction
mediapipe>=0.10.0

# Optional: Web UI (uncomment if needed)
# streamlit>=1.20.0
# gradio>=3.0.0

# Development and Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Reproducibility
setproctitle>=1.2.0

